requirements.txt
absl-py==1.4.0
altgraph==0.17.4
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astor==0.8.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.20
django-import-export==3.2.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fake-useragent==1.1.3
Faker==18.13.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.38.0
gast==0.5.4
grpcio==1.56.0
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
interchange==2021.0.4
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.9.3
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.1
mock==5.0.2
monotonic==1.6
mysqlclient @ file:///D:/Venvs/env_temp/Scripts/mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl#sha256=683269b350813b3d593e290e0d565b0de6bcd7f8b3c5cec3025260b3dfcd3b57
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==1.1.5
pandocfilters==1.5.0
pansi==2020.7.3
parse==1.19.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.5
py2neo==2021.2.3
pycparser==2.21
pyee==8.2.2
Pygments==2.15.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.5
pyparsing==3.1.0
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-html==0.10.0
scipy==1.7.3
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.4.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==1.26.16
w3lib==2.1.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
